requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
openai>=1.40.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0

# Web interface
gradio>=4.40.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
tiktoken>=0.7.0

# Utilities
tqdm>=4.66.0
pydantic>=2.8.0